Added First Encounter
"Jerry the Slime" has been added!
Jerry is a basic enemy who only wanders and spreads only only basic bullet pattern. Easy enough for the player to get a feel for the combat.
With the first encounter being added, this also meant adding player shooting and a health bar for both the player and enemy.